Modification register ajout de confirm password && edit password fonction placé dans user controller
